Flare
Flare1:Flare normal
Flare2:blinking
NOTE:
Flare cannot be destroyed by jumping on and can only be destroyed by a
fireball or falling in lava so on and so forth
NOTE2:
Ambers and the other mob (I forgot the name) can be killed with a
fireball and by jumping on. The other mob will have 2 hearts so you'll
have to use 2 attacks.